pyproject.toml
[project]
name = "backpack-grid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.5",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "websocket-client>=1.6.4",
]
